+++ src/CommandPalette.h
@@ -7,7 +7,8 @@ constexpr const char* kPalettePrefixCommands = ">";
 constexpr const char* kPalettePrefixFileHistory = "#";
 constexpr const char* kPalettePrefixTabs = "@";
 constexpr const char* kPalettePrefixEverything = ":";
-constexpr const char* kPalettePrefixTOC = "*";
+constexpr const char* kPalettePrefixTOC = "%";
+constexpr const char* kPalettePrefixTOCLegacy = "*";
 constexpr const char* kPalettePrefixFavorites = "$";
 
 void RunCommandPalette(MainWindow*, Str prefix, int smartTabAdvance);